moving mapboxtilesetcreate to a static class to add testing
